# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StayBnB Room Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List rooms with filters
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [ENTIRE, PRIVATE, SHARED]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: keyword
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomListResponse"

    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new room (HOST only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomRequest"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "403":
          description: Not a host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room with reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetail"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateRoom
      tags: [rooms]
      summary: Update room (HOST/owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomRequest"
      responses:
        "200":
          description: Room updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "403":
          description: Not owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete room (HOST/owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Room deleted
        "403":
          description: Not owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms/host/my:
    get:
      operationId: getHostRooms
      tags: [rooms]
      summary: Get current host's rooms
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Host's rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
  /rooms/{id}/reviews:
    get:
      operationId: getRoomReviews
      tags: [reviews]
      summary: Get reviews for a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review for a room
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequest"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "403":
          description: No completed booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reviews/{id}:
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review (admin or author)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Review deleted
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [GUEST, HOST, ADMIN]
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        profilePicture:
          type: string
          nullable: true
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isVerified
        - createdAt
...
    Room:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [ENTIRE, PRIVATE, SHARED]
        pricePerNight:
          type: number
        location:
          type: string
        city:
          type: string
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
        avgRating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        host:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - type
        - pricePerNight
        - location
        - city
        - amenities
        - images
        - isAvailable
        - reviewCount
        - host
        - createdAt
    RoomDetail:
      allOf:
        - $ref: "#/components/schemas/Room"
        - type: object
          properties:
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
          required:
            - reviews
    RoomListResponse:
      type: object
      properties:
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/Room"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - rooms
        - total
        - page
        - limit
        - totalPages
    RoomRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 5
        description:
          type: string
          minLength: 10
        type:
          type: string
          enum: [ENTIRE, PRIVATE, SHARED]
        pricePerNight:
          type: number
          minimum: 1
        location:
          type: string
        city:
          type: string
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
      required:
        - title
        - description
        - type
        - pricePerNight
        - location
        - city
        - amenities
        - images
...
    Review:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - roomId
        - user
        - rating
        - comment
        - createdAt
    ReviewRequest:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          minLength: 5
      required:
        - rating
        - comment
...