# lib/api-zod/src/generated/types/reviewRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StayBnB Room Booking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewRequest {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  /** @minLength 5 */
  comment: string;
}