# lib/api-zod/src/generated/types/roomListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StayBnB Room Booking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Room } from "./room";

export interface RoomListResponse {
  rooms: Room[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}